THOOOMMMASSSS is stoopppiiiddd

+++ Chapter 9/Example.java
@@ -3,7 +3,7 @@ public class Example
     public static void main ( String[] args)
     {
         int x = 2; 
-        long hoursWorked = x;
+        double hoursWorked = x;
         double payRate =10, taxRate = .634;
         System.out.println("Hours Worked:" + hoursWorked );
         System.out.println("pay Amount: " + (hoursWorked * payRate) );
